petlog/backend/app/db.py
# backend/app/db.py
from __future__ import annotations

import os
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker, DeclarativeBase

# ✅ 若你有用 .env
try:
    from dotenv import load_dotenv
    load_dotenv()
except Exception:
    pass

DATABASE_URL = os.getenv("DATABASE_URL")
if not DATABASE_URL:
    raise RuntimeError("DATABASE_URL is not set. Please set it in .env")



connect_args = {}
# Zeabur often uses a direct connection inside the cluster or specific SSL settings.
# Trying 'prefer' or removing sslmode if 'require' fails.
if "postgresql" in DATABASE_URL:
    connect_args = {"sslmode": "prefer"}

# Adjust for sqlite
if "sqlite" in DATABASE_URL:
    connect_args = {"check_same_thread": False}

engine = create_engine(DATABASE_URL, pool_pre_ping=True, connect_args=connect_args)

SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)


class Base(DeclarativeBase):
    pass


def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
